refactor: separate scan job with different reconciler
