+++ pkg/vulnerabilityreport/controller/scanjob.go
@@ -1,0 +1,321 @@
+package controller
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/aquasecurity/trivy-operator/pkg/apis/aquasecurity/v1alpha1"
+	"github.com/aquasecurity/trivy-operator/pkg/exposedsecretreport"
+	"github.com/aquasecurity/trivy-operator/pkg/kube"
+	"github.com/aquasecurity/trivy-operator/pkg/operator/etc"
+	. "github.com/aquasecurity/trivy-operator/pkg/operator/predicate"
+	"github.com/aquasecurity/trivy-operator/pkg/trivyoperator"
+	"github.com/aquasecurity/trivy-operator/pkg/vulnerabilityreport"
+	"github.com/go-logr/logr"
+	"go.uber.org/multierr"
+	batchv1 "k8s.io/api/batch/v1"
+	k8sapierror "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/builder"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/predicate"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+)
+
+// WorkloadController watches Kubernetes workloads and generates
+// v1alpha1.VulnerabilityReport instances using vulnerability scanner that that
+// implements the Plugin interface.
+type ScanJobController struct {
+	logr.Logger
+	etc.Config
+	kube.ObjectResolver
+	kube.LogsReader
+	vulnerabilityreport.Plugin
+	trivyoperator.PluginContext
+	trivyoperator.ConfigData
+	VulnerabilityReadWriter vulnerabilityreport.ReadWriter
+	ExposedSecretReadWriter exposedsecretreport.ReadWriter
+}
+
+// Manage scan jobs with image pull secrets
+// kubebuilder:rbac:groups="",resources=secrets,verbs=create;update
+//+kubebuilder:rbac:groups=batch,resources=jobs,verbs=get;list;watch;create;delete
+
+func (r *ScanJobController) SetupWithManager(mgr ctrl.Manager) error {
+	var predicates []predicate.Predicate
+	if !r.ConfigData.VulnerabilityScanJobsInSameNamespace() {
+		predicates = append(predicates, InNamespace(r.Config.Namespace))
+	}
+	predicates = append(predicates, ManagedByTrivyOperator, IsVulnerabilityReportScan, JobHasAnyCondition)
+	return ctrl.NewControllerManagedBy(mgr).
+		For(&batchv1.Job{}, builder.WithPredicates(predicates...)).
+		Complete(r.reconcileJobs())
+}
+
+func (r *ScanJobController) hasReports(ctx context.Context, owner kube.ObjectRef, hash string, images kube.ContainerImages) (bool, error) {
+	hasVulnerabilityReports, err := r.hasVulnerabilityReports(ctx, owner, hash, images)
+	if err != nil {
+		return false, err
+	}
+
+	hasSecretReports, err := r.hasSecretReports(ctx, owner, hash, images)
+	if err != nil {
+		return false, err
+	}
+
+	return hasVulnerabilityReports && hasSecretReports, nil
+}
+
+func (r *ScanJobController) hasVulnerabilityReports(ctx context.Context, owner kube.ObjectRef, hash string, images kube.ContainerImages) (bool, error) {
+	// TODO FindByOwner should accept optional label selector to further narrow down search results
+	list, err := r.VulnerabilityReadWriter.FindByOwner(ctx, owner)
+	if err != nil {
+		return false, err
+	}
+
+	actual := map[string]bool{}
+	for _, report := range list {
+		if containerName, ok := report.Labels[trivyoperator.LabelContainerName]; ok {
+			if hash == report.Labels[trivyoperator.LabelResourceSpecHash] {
+				actual[containerName] = true
+			}
+		}
+	}
+
+	return compareReports(actual, images), nil
+}
+
+func (r *ScanJobController) hasSecretReports(ctx context.Context, owner kube.ObjectRef, hash string, images kube.ContainerImages) (bool, error) {
+	// TODO FindByOwner should accept optional label selector to further narrow down search results
+	list, err := r.ExposedSecretReadWriter.FindByOwner(ctx, owner)
+	if err != nil {
+		return false, err
+	}
+
+	actual := map[string]bool{}
+	for _, report := range list {
+		if containerName, ok := report.Labels[trivyoperator.LabelContainerName]; ok {
+			if hash == report.Labels[trivyoperator.LabelResourceSpecHash] {
+				actual[containerName] = true
+			}
+		}
+	}
+
+	return compareReports(actual, images), nil
+}
+
+func (r *ScanJobController) reconcileJobs() reconcile.Func {
+	return func(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
+		log := r.Logger.WithValues("job", req.NamespacedName)
+
+		job := &batchv1.Job{}
+		err := r.Client.Get(ctx, req.NamespacedName, job)
+		if err != nil {
+			if k8sapierror.IsNotFound(err) {
+				log.V(1).Info("Ignoring cached job that must have been deleted")
+				return ctrl.Result{}, nil
+			}
+			return ctrl.Result{}, fmt.Errorf("getting job from cache: %w", err)
+		}
+
+		if len(job.Status.Conditions) == 0 {
+			log.V(1).Info("Ignoring Job without conditions")
+			return ctrl.Result{}, nil
+		}
+
+		switch jobCondition := job.Status.Conditions[0].Type; jobCondition {
+		case batchv1.JobComplete:
+			err = r.processCompleteScanJob(ctx, job)
+		case batchv1.JobFailed:
+			err = r.processFailedScanJob(ctx, job)
+		default:
+			err = fmt.Errorf("unrecognized scan job condition: %v", jobCondition)
+		}
+
+		return ctrl.Result{}, err
+	}
+
+}
+
+func (r *ScanJobController) processCompleteScanJob(ctx context.Context, job *batchv1.Job) error {
+	log := r.Logger.WithValues("job", fmt.Sprintf("%s/%s", job.Namespace, job.Name))
+
+	ownerRef, err := kube.ObjectRefFromObjectMeta(job.ObjectMeta)
+	if err != nil {
+		return fmt.Errorf("getting owner ref from scan job metadata: %w", err)
+	}
+
+	owner, err := r.ObjectFromObjectRef(ctx, ownerRef)
+	if err != nil {
+		if k8sapierror.IsNotFound(err) {
+			log.V(1).Info("Report owner must have been deleted", "owner", owner)
+			return r.deleteJob(ctx, job)
+		}
+		return fmt.Errorf("getting object from object ref: %w", err)
+	}
+
+	containerImages, err := kube.GetContainerImagesFromJob(job)
+	if err != nil {
+		return fmt.Errorf("getting container images: %w", err)
+	}
+
+	podSpecHash, ok := job.Labels[trivyoperator.LabelResourceSpecHash]
+	if !ok {
+		return fmt.Errorf("expected label %s not set", trivyoperator.LabelResourceSpecHash)
+	}
+
+	hasReports, err := r.hasReports(ctx, ownerRef, podSpecHash, containerImages)
+	if err != nil {
+		return err
+	}
+
+	if hasReports {
+		log.V(1).Info("VulnerabilityReports already exist", "owner", owner)
+		log.V(1).Info("Deleting complete scan job", "owner", owner)
+		return r.deleteJob(ctx, job)
+	}
+
+	var vulnerabilityReports []v1alpha1.VulnerabilityReport
+	var secretReports []v1alpha1.ExposedSecretReport
+
+	var merr error
+	for containerName, containerImage := range containerImages {
+		vulnReports, secReports, err := r.processScanJobResults(ctx, job, containerName, containerImage, owner)
+		if err != nil {
+			merr = multierr.Append(merr, err)
+		}
+		vulnerabilityReports = append(vulnerabilityReports, vulnReports...)
+		secretReports = append(secretReports, secReports...)
+	}
+	if merr != nil {
+		return merr
+	}
+
+	if r.Config.VulnerabilityScannerEnabled {
+		err = r.VulnerabilityReadWriter.Write(ctx, vulnerabilityReports)
+		if err != nil {
+			return err
+		}
+	}
+
+	if r.Config.ExposedSecretScannerEnabled {
+		err = r.ExposedSecretReadWriter.Write(ctx, secretReports)
+		if err != nil {
+			return err
+		}
+	}
+
+	log.V(1).Info("Deleting complete scan job", "owner", owner)
+	return r.deleteJob(ctx, job)
+}
+
+func (r *ScanJobController) processScanJobResults(ctx context.Context, job *batchv1.Job, containerName, containerImage string, owner client.Object) ([]v1alpha1.VulnerabilityReport, []v1alpha1.ExposedSecretReport, error) {
+	log := r.Logger.WithValues("job-results-processor", fmt.Sprintf("%s/%s", job.Namespace, job.Name))
+
+	var vulnerabilityReports []v1alpha1.VulnerabilityReport
+	var secretReports []v1alpha1.ExposedSecretReport
+
+	podSpecHash, ok := job.Labels[trivyoperator.LabelResourceSpecHash]
+	if !ok {
+		return nil, nil, fmt.Errorf("expected label %s not set", trivyoperator.LabelResourceSpecHash)
+	}
+
+	logsStream, err := r.LogsReader.GetLogsByJobAndContainerName(ctx, job, containerName)
+	if err != nil {
+		if k8sapierror.IsNotFound(err) {
+			log.V(1).Info("Cached job must have been deleted")
+			return nil, nil, nil
+		}
+		if kube.IsPodControlledByJobNotFound(err) {
+			log.V(1).Info("Pod must have been deleted")
+			return nil, nil, r.deleteJob(ctx, job)
+		}
+		return nil, nil, fmt.Errorf("getting logs for pod %q: %w", job.Namespace+"/"+job.Name, err)
+	}
+
+	defer func() {
+		err := logsStream.Close()
+		if err != nil {
+			log.V(1).Error(err, "could not close log stream")
+		}
+	}()
+
+	vulnReportData, secretReportData, err := r.Plugin.ParseReportData(r.PluginContext, containerImage, logsStream)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	resourceLabelsToInclude := r.GetReportResourceLabels()
+
+	reportBuilder := vulnerabilityreport.NewReportBuilder(r.Client.Scheme()).
+		Controller(owner).
+		Container(containerName).
+		Data(vulnReportData).
+		PodSpecHash(podSpecHash).
+		ResourceLabelsToInclude(resourceLabelsToInclude)
+
+	if r.Config.ScannerReportTTL != nil {
+		reportBuilder.ReportTTL(r.Config.ScannerReportTTL)
+	}
+
+	report, err := reportBuilder.Get()
+	if err != nil {
+		return nil, nil, err
+	}
+
+	secretReportBuilder := exposedsecretreport.NewReportBuilder(r.Client.Scheme()).
+		Controller(owner).
+		Container(containerName).
+		Data(secretReportData).
+		PodSpecHash(podSpecHash).
+		ResourceLabelsToInclude(resourceLabelsToInclude)
+	if r.Config.ScannerReportTTL != nil {
+		secretReportBuilder.ReportTTL(r.Config.ScannerReportTTL)
+	}
+	secretReport, err := secretReportBuilder.Get()
+	if err != nil {
+		return nil, nil, err
+	}
+
+	vulnerabilityReports = append(vulnerabilityReports, report)
+	secretReports = append(secretReports, secretReport)
+
+	return vulnerabilityReports, secretReports, nil
+}
+
+func (r *ScanJobController) processFailedScanJob(ctx context.Context, scanJob *batchv1.Job) error {
+	log := r.Logger.WithValues("job", fmt.Sprintf("%s/%s", scanJob.Namespace, scanJob.Name))
+
+	statuses, err := r.GetTerminatedContainersStatusesByJob(ctx, scanJob)
+	if err != nil {
+		if k8sapierror.IsNotFound(err) {
+			log.V(1).Info("Cached job must have been deleted")
+			return nil
+		}
+		if kube.IsPodControlledByJobNotFound(err) {
+			log.V(1).Info("Pod must have been deleted")
+			return r.deleteJob(ctx, scanJob)
+		}
+		return err
+	}
+	for container, status := range statuses {
+		if status.ExitCode == 0 {
+			continue
+		}
+		log.Error(nil, "Scan job container", "container", container, "status.reason", status.Reason, "status.message", status.Message)
+	}
+	log.V(1).Info("Deleting failed scan job")
+	return r.deleteJob(ctx, scanJob)
+}
+
+func (r *ScanJobController) deleteJob(ctx context.Context, job *batchv1.Job) error {
+	err := r.Client.Delete(ctx, job, client.PropagationPolicy(metav1.DeletePropagationBackground))
+	if err != nil {
+		if k8sapierror.IsNotFound(err) {
+			return nil
+		}
+		return fmt.Errorf("deleting job: %w", err)
+	}
+	return nil
+}
